Recompute the pairwise alignment from the branch we are leaving.
This is problematic, because its not clear how we should order
unordered columns here.

If we knew how to order unordered columns, we could compute this
alignment by composing the pairwise alignments of the branches
we are leaving.

It is also unclear how one should treat unordered columns when
composing pairwise alignments -- do the pairwise alignments only
mean something on a given branch?  Or, does composing them make
sense?

For TKF1/2 composing them seems to make sense, at least if we
know the DIRECTION of the branches...